pyproject.toml
[tool.poetry]
name = "repl_nix_notion-qa"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
faiss-cpu = "^1.7.4"
streamlit = "^1.22.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
